pyproject.toml
[project]
name = "python-decouple-openai-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.3",
    "python-decouple>=3.8",
    "rich>=14.0.0",
]
